refactored some of the sections into seperate components
